Implement TcpMonitor as TcpProcessor and move to monitor package
